requirements.txt

APScheduler==3.11.0
fastapi==0.115.11
passlib==1.7.4
psycopg2==2.9.10
pydantic==2.10.6
python-dotenv==1.0.1
python-jose==3.4.0
redis==5.2.1
shortuuid==1.0.13
SQLAlchemy==2.0.39
uvicorn==0.34.0
pydantic[email]
python-multipart==0.0.20